changed ExtensionsStatusBadge-new to ExtensionsStatusBadge
