+++ src/components/common/ColonyExtensions/ExtensionCard/ExtensionCard.tsx
@@ -3,7 +3,7 @@ import { FormattedMessage } from 'react-intl';
 import { useParams } from 'react-router-dom';
 
 import Card from '~shared/Card';
-import ExtensionStatusBadge from '~common/Extensions/ExtensionStatusBadge';
+import ExtensionStatusBadge from '~common/ExtensionStatusBadge';
 import { Heading4 } from '~shared/Heading';
 import Icon from '~shared/Icon';
 import Link from '~shared/Link';
