+++ src/components/common/Extensions/SubNavigation/partials/DropdownContent/PayDropdown.tsx
@@ -2,7 +2,7 @@ import React, { PropsWithChildren, FC } from 'react';
 import { FormattedMessage } from 'react-intl';
 import Button from '~shared/Extensions/Button/Button';
 import { LEARN_MORE_PAYMENTS } from '~constants';
-import ExtensionStatusBadge from '~common/Extensions/ExtensionStatusBadge-new';
+import ExtensionStatusBadge from '~common/Extensions/ExtensionStatusBadge';
 import LinkItem from '../LinkItem';
 import styles from './DropdownContent.module.css';
 import { MSG } from './consts';
