+++ src/components/shared/Extensions/Fields/RadioList/RadioBase.tsx
@@ -3,7 +3,7 @@ import clsx from 'clsx';
 import { useIntl } from 'react-intl';
 import styles from './RadioList.module.css';
 import { RadioBaseProps } from './types';
-import ExtensionStatusBadge from '~common/Extensions/ExtensionStatusBadge-new';
+import ExtensionStatusBadge from '~common/Extensions/ExtensionStatusBadge';
 import Tooltip from '~shared/Extensions/Tooltip/Tooltip';
 import Icon from '~shared/Icon';
